Fix circular deps.
